Someting changed

+++ backup.sh
@@ -21,12 +21,14 @@ for inner_fpath in $(find "$root" -mindepth 2 -type f -not -path "*.git/*"); do
     #     continue
     # fi
 
-    # copy files from system to git repo
-    if [ "$(stat -c '%U' "$out_fpath")" != "$USER" ]; then
-	# if file owned not by currect user, copy it with sudo
-	sudo cp "$out_fpath" "$inner_fpath"
-    else
-        cp "$out_fpath" "$inner_fpath"
+    if [[ -f "$out_fpath" ]]; then
+    	# copy files from system to git repo
+    	if [ "$(stat -c '%U' "$out_fpath")" != "$USER" ]; then
+    	    # if file owned not by currect user, copy it with sudo
+    	    sudo cp "$out_fpath" "$inner_fpath"
+    	else
+    	    cp "$out_fpath" "$inner_fpath"
+    	fi
     fi
 done
